+++ src/main/java/miyucomics/hexical/casting/patterns/getters/OpGetLivingEntityData.kt
@@ -3,10 +3,7 @@ package miyucomics.hexical.casting.patterns.getters
 import at.petrak.hexcasting.api.spell.ConstMediaAction
 import at.petrak.hexcasting.api.spell.casting.CastingContext
 import at.petrak.hexcasting.api.spell.getLivingEntityButNotArmorStand
-import at.petrak.hexcasting.api.spell.iota.BooleanIota
-import at.petrak.hexcasting.api.spell.iota.Iota
-import at.petrak.hexcasting.api.spell.iota.ListIota
-import at.petrak.hexcasting.api.spell.iota.NullIota
+import at.petrak.hexcasting.api.spell.iota.*
 import miyucomics.hexical.casting.iota.IdentifierIota
 import net.minecraft.util.registry.Registry
 
@@ -16,9 +13,13 @@ class OpGetLivingEntityData(private val mode: Int) : ConstMediaAction {
 		val entity = args.getLivingEntityButNotArmorStand(0, argc)
 		return listOf(
 			when (mode) {
-				0 -> BooleanIota(entity.isSleeping)
-				1 -> BooleanIota(entity.isSprinting)
-				2 -> {
+				0 -> DoubleIota(entity.health.toDouble())
+				1 -> DoubleIota(entity.maxHealth.toDouble())
+				2 -> DoubleIota(entity.air.toDouble() / 20 + 1)
+				3 -> DoubleIota(entity.maxAir.toDouble() / 20 + 1)
+				4 -> BooleanIota(entity.isSleeping)
+				5 -> BooleanIota(entity.isSprinting)
+				6 -> {
 					val list = mutableListOf<Iota>()
 					for (effect in entity.statusEffects)
 						list.add(IdentifierIota(Registry.STATUS_EFFECT.getId(effect.effectType)!!))