+++ src/main/java/miyucomics/hexical/casting/patterns/getters/OpGetPositionData.kt
@@ -1,0 +1,24 @@
+package miyucomics.hexical.casting.patterns.getters
+
+import at.petrak.hexcasting.api.spell.ConstMediaAction
+import at.petrak.hexcasting.api.spell.casting.CastingContext
+import at.petrak.hexcasting.api.spell.getBlockPos
+import at.petrak.hexcasting.api.spell.iota.DoubleIota
+import at.petrak.hexcasting.api.spell.iota.Iota
+import at.petrak.hexcasting.api.spell.iota.NullIota
+import miyucomics.hexical.casting.iota.IdentifierIota
+
+class OpGetPositionData(private val mode: Int) : ConstMediaAction {
+	override val argc = 1
+	override fun execute(args: List<Iota>, ctx: CastingContext): List<Iota> {
+		val position = args.getBlockPos(0, argc)
+		ctx.assertVecInRange(position)
+		return listOf(
+			when (mode) {
+				0 -> DoubleIota(ctx.world.getLightLevel(position).toDouble())
+				1 -> IdentifierIota(ctx.world.getBiome(position).key.get().value)
+				else -> NullIota()
+			}
+		)
+	}
+}